Update Academy course to use current HoloToolkit
Modifies the project to utilize the new HoloToolkit input models for gestures and speech as well as build in Unity's 2017.2 Beta release.

The academy course isn't entirely complete and will need some more flushing out:

- Keywords are hard-coded; current HoloToolKit doesn't allow assigning keyword handler in the UI (as it did previously)

-TapToPlace from toolkit doesn't have "pickup" and "place" sound effects

-TopToPlace only allows selection on Hub's Base and not the "Orb"

-Keyword recognizer doesn't seem to be working very well; need to debug

The HoloToolkit contained in this project is just a placeholder, using a recent snap of the code, and should be replaced with the "finished" update for Mixed Reality.